correct ordering, readme, props renamed for clarity.
